tambahin filter team
Now we can start to see the benefit of the work we did in the last section.
We can add filtering by a new option with just a few simple changes to the model.

We updated the model to add team_id to the list of valid FILTER_PARAMS,
added the by_team scope, and then added by_team to the filters method.

No need to change our controller or touch the Filterable module —
updating the model is all we need.